polling/requirements.txt
requests
google-python-cloud-debugger
google-cloud-logging